+++ index.html
@@ -8,6 +8,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script>src = "functions.js"</script>
   <script>src = "main.js"</script>
+  <script type="module" src="main.js"></script>
   <title>Кекстаграм</title>
 </head>
