+++ js/util.js
@@ -1,0 +1,3 @@
+export function getRandomInt(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
